feat(sankey): adjust labels for vertical layout
