+++ packages/sankey/src/SankeyLabels.js
@@ -14,16 +14,14 @@ import { colorMotionSpring, getInterpolatedColor, motionPropTypes } from '@nivo/
 
 const SankeyLabels = ({
     nodes,
-
+    layout,
     width,
-
+    height,
     labelPosition,
     labelPadding,
     labelOrientation,
     getLabelTextColor,
-
     theme,
-
     animate,
     motionDamping,
     motionStiffness,
@@ -31,22 +29,45 @@ const SankeyLabels = ({
     const labelRotation = labelOrientation === 'vertical' ? -90 : 0
     const labels = nodes.map(node => {
         let x
+        let y
         let textAnchor
-        if (node.x < width / 2) {
-            if (labelPosition === 'inside') {
-                x = node.x1 + labelPadding
-                textAnchor = labelOrientation === 'vertical' ? 'middle' : 'start'
+        if (layout === 'horizontal') {
+            y = node.y + node.height / 2
+            if (node.x < width / 2) {
+                if (labelPosition === 'inside') {
+                    x = node.x1 + labelPadding
+                    textAnchor = labelOrientation === 'vertical' ? 'middle' : 'start'
+                } else {
+                    x = node.x - labelPadding
+                    textAnchor = labelOrientation === 'vertical' ? 'middle' : 'end'
+                }
             } else {
-                x = node.x - labelPadding
-                textAnchor = labelOrientation === 'vertical' ? 'middle' : 'end'
+                if (labelPosition === 'inside') {
+                    x = node.x - labelPadding
+                    textAnchor = labelOrientation === 'vertical' ? 'middle' : 'end'
+                } else {
+                    x = node.x1 + labelPadding
+                    textAnchor = labelOrientation === 'vertical' ? 'middle' : 'start'
+                }
             }
-        } else {
-            if (labelPosition === 'inside') {
-                x = node.x - labelPadding
-                textAnchor = labelOrientation === 'vertical' ? 'middle' : 'end'
+        } else if (layout === 'vertical') {
+            x = node.x + node.width / 2
+            if (node.y < height / 2) {
+                if (labelPosition === 'inside') {
+                    y = node.y1 + labelPadding
+                    textAnchor = labelOrientation === 'vertical' ? 'end' : 'middle'
+                } else {
+                    y = node.y - labelPadding
+                    textAnchor = labelOrientation === 'vertical' ? 'start' : 'middle'
+                }
             } else {
-                x = node.x1 + labelPadding
-                textAnchor = labelOrientation === 'vertical' ? 'middle' : 'start'
+                if (labelPosition === 'inside') {
+                    y = node.y - labelPadding
+                    textAnchor = labelOrientation === 'vertical' ? 'start' : 'middle'
+                } else {
+                    y = node.y1 + labelPadding
+                    textAnchor = labelOrientation === 'vertical' ? 'end' : 'middle'
+                }
             }
         }
 
@@ -54,7 +75,7 @@ const SankeyLabels = ({
             id: node.id,
             label: node.label,
             x,
-            y: node.y + node.height / 2,
+            y,
             textAnchor,
             color: getLabelTextColor(node),
         }
@@ -144,9 +165,9 @@ SankeyLabels.propTypes = {
             height: PropTypes.number.isRequired,
         })
     ).isRequired,
-
+    layout: PropTypes.oneOf(['horizontal', 'vertical']).isRequired,
     width: PropTypes.number.isRequired,
-
+    height: PropTypes.number.isRequired,
     labelPosition: PropTypes.oneOf(['inside', 'outside']).isRequired,
     labelPadding: PropTypes.number.isRequired,
     labelOrientation: PropTypes.oneOf(['horizontal', 'vertical']).isRequired,
